fix: handle async replication changed during initialization

fix: stoppable hashtree initialization


fix: proper unlock after initialization


fix: wait for hashbeater to stop


fix: do not trigger shard loading to avoid potential races


fix: shard drop race


fix: shard drop race


chore: set explicit context with timeout


chore: revert unlock if error for now


chore: remove context with explicit timeout


chore: set explicit context with timeout


chore: reduce time fuse delay


fix: simpler difference collection


chore: improve method signature


chore: improve method signature


fix: unlock waiting for hashbeater to finish


chore: improve error handling


fix: logging when no diff was found


chore: simplified async stop


fix: do not attempt to fetch secondary if parsing returns error


fix: init reusable node with secondary keys
