+++ adapters/repos/db/shard_backup.go
@@ -22,7 +22,7 @@ import (
 )
 
 // HaltForTransfer stops compaction, and flushing memtable and commit log to begin with backup or cloud offload
-func (s *Shard) HaltForTransfer(ctx context.Context) (err error) {
+func (s *Shard) HaltForTransfer(ctx context.Context, offloading bool) (err error) {
 	defer func() {
 		if err != nil {
 			err = fmt.Errorf("pause compaction: %w", err)
@@ -32,13 +32,11 @@ func (s *Shard) HaltForTransfer(ctx context.Context) (err error) {
 		}
 	}()
 
-	// NOTE: async replication may be paused during backup
-	// only if the hashtree can be resumed but object updates
-	// or deletions requires special handling.
-	// Curently hashtree is not included into the backup files,
-	// the hashtree will be automatically regenerated when restoring
-	// s.mayStopHashBeater()
-	// s.mayCloseHashTree()
+	if offloading {
+		// TODO: tenant offloading is calling HaltForTransfer but
+		// if Shutdown is called this step is not needed
+		s.mayStopAsyncReplication()
+	}
 
 	if err = s.store.PauseCompaction(ctx); err != nil {
 		return fmt.Errorf("pause compaction: %w", err)
@@ -111,10 +109,6 @@ func (s *Shard) ListBackupFiles(ctx context.Context, ret *backup.ShardDescriptor
 func (s *Shard) resumeMaintenanceCycles(ctx context.Context) error {
 	g := enterrors.NewErrorGroupWrapper(s.index.logger)
 
-	// NOTE: async replication may be resumed if paused in HaltForTransfer method
-	//g.Go(func() error {
-	//	return s.UpdateAsyncReplication(ctx, s.index.asyncReplicationEnabled())
-	//})
 	g.Go(func() error {
 		return s.store.ResumeCompaction(ctx)
 	})
