+++ adapters/repos/db/shard_drop.go
@@ -42,15 +42,16 @@ func (s *Shard) drop() (err error) {
 		s.clearDimensionMetrics()
 	}
 
-	s.mayStopHashBeater()
-	s.mayCleanupHashTree()
+	s.index.logger.WithFields(logrus.Fields{
+		"action": "drop_shard",
+		"class":  s.class.Class,
+		"shard":  s.name,
+	}).Debug("dropping shard")
+
+	s.mayStopAsyncReplication()
 
 	ctx, cancel := context.WithTimeout(context.TODO(), 20*time.Second)
 	defer cancel()
-	s.index.logger.WithFields(logrus.Fields{
-		"action":   "drop_shard",
-		"duration": 5 * time.Second,
-	}).Debug("context.WithTimeout")
 
 	// unregister all callbacks at once, in parallel
 	if err = cyclemanager.NewCombinedCallbackCtrl(0, s.index.logger,
@@ -127,6 +128,12 @@ func (s *Shard) drop() (err error) {
 	}
 
 	s.metrics.baseMetrics.FinishUnloadingShard(s.index.Config.ClassName.String())
+
+	s.index.logger.WithFields(logrus.Fields{
+		"action": "drop_shard",
+		"class":  s.class.Class,
+		"shard":  s.name,
+	}).Debug("shard successfully dropped")
 
 	return nil
 }
