+++ adapters/repos/db/shard_shutdown.go
@@ -63,8 +63,7 @@ func (s *Shard) Shutdown(ctx context.Context) (err error) {
 	).Unregister(ctx)
 	ec.Add(err)
 
-	s.mayStopHashBeater()
-	s.mayCloseAndStoreHashTree()
+	s.mayStopAsyncReplication()
 
 	if s.hasTargetVectors() {
 		// TODO run in parallel?
