+++ test/acceptance/replication/async_replication/async_repair_multi_tenancy_test.go
@@ -70,7 +70,7 @@ func (suite *AsyncReplicationTestSuite) TestAsyncRepairMultiTenancyScenario() {
 		}
 	}()
 
-	ctx, cancel := context.WithTimeout(mainCtx, 10*time.Minute)
+	ctx, cancel := context.WithTimeout(mainCtx, 15*time.Minute)
 	defer cancel()
 
 	paragraphClass := articles.ParagraphsClass()
